Revert "Move Lloyd-Max fixtures into owning tests"
This reverts commit 3ec6dd25e2d632b179c5e19ccfaee9ae03b6bc43.
